Puzzle 1 switch patch and new tile id system
This new tile id system should work as it is based off the puzzle number.
I haven't tested with anything other than 8x8 but hopefully, we can probably add n to the key if its still buggy.
The the puzzle switching bug has been fully patched, realized I want using puzzle 1s preset and altering it by reference
so when you go from 2 or 3 to 1 it would still be 2 or 3, now that is fixed as i just copy the values instead of referencing it.